Introduced Id types like TaskId and ResultId.

+++ src/libclient/scheduler/schedulermodel.cpp
@@ -181,7 +181,7 @@ QVariant SchedulerModel::data(const QModelIndex &index, int role) const
         return testDefinition.timing()->type();
 
     case IdRole:
-        return testDefinition.id();
+        return testDefinition.id().toInt();
     }
 
     return QVariant();
